items/actions/general: move assault course handlers
