+++ src/libtrx/game/items/actions/common.c
@@ -3,8 +3,6 @@
 
 static void (*m_Routines[ITEM_ACTION_NUMBER_OF])(ITEM *item) = {};
 
-extern void Item_ActionRunLegacy(ITEM_TRX_ACTION action_id, ITEM *item);
-
 void Item_ActionRegister(
     const ITEM_TRX_ACTION action, void (*const action_func)(ITEM *item))
 {
@@ -16,9 +14,6 @@ void Item_ActionRun(const ITEM_TRX_ACTION action_id, ITEM *const item)
     if (action_id >= 0 && action_id < ITEM_ACTION_NUMBER_OF
         && m_Routines[action_id] != nullptr) {
         m_Routines[action_id](item);
-    } else {
-        // TODO: remove once all routines merged
-        Item_ActionRunLegacy(action_id, item);
     }
 }
