--- src/tr1/game/item_actions.c
@@ -1,5 +1,0 @@
-#include <libtrx/game/items.h>
-
-void Item_ActionRunLegacy(ITEM_TRX_ACTION action_id, ITEM *item)
-{
-}
