+++ src/tr2/meson.build
@@ -105,7 +105,6 @@ sources = [
   'game/inventory.c',
   'game/inventory_ring/control.c',
   'game/inventory_ring/vars.c',
-  'game/item_actions.c',
   'game/lara/common.c',
   'game/lara/flare.c',
   'game/lara/misc.c',
